Smart_ATS_App_Codebase/requirements.txt
streamlit
pyrebase4
python-dotenv
PyPDF2
google-generativeai
pandas
